backend/requirements.txt
dependency-injector
fastapi
uvicorn
pyyaml
sqlalchemy
pytest
requests
pytest-cov
pydantic
factory-boy